#88 [Refactor] 이미지 스크린샷, 공유기능 utils로 분리
